requirements.txt
adversarial-robustness-toolbox==1.17.1
anaconda-anon-usage==0.4.4
archspec==0.2.3
astunparse==1.6.3
black==24.4.2
brotli==1.0.9
conda-build==24.3.0
conda-content-trust==0.2.0
conda-libmamba-solver==23.7.0
distro==1.8.0
expecttest==0.2.1
fqdn==1.5.1
gmpy2==2.1.2
hydra-core==1.3.2
hypothesis==6.100.1
isoduration==20.11.0
jupyter==1.0.0
lark==1.1.9
mkl-fft==1.3.8
mkl-random==1.2.4
multiprocess==0.70.16
optree==0.11.0
pip-chill==1.0.3
pysocks==1.7.1
scikit-image==0.23.2
torchaudio==2.3.0
torchelastic==0.2.2
torchvision==0.18.0
triton==2.3.0
truststore==0.8.0
types-dataclasses==0.6.6
uri-template==1.3.0
webcolors==1.13